+++ packages/lumina/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @prism-lang/lumina
 
+## 0.1.2
+
+### Patch Changes
+
+- @prism-lang/confidence@1.2.6
+
 ## 0.1.1
 
 ### Patch Changes